fixing warnings, refs #3
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@19650 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/traffic_lights/MSPhasedTrafficLightLogic.cpp
@@ -127,7 +127,7 @@ MSPhasedTrafficLightLogic::getPhases() const {
 
 const MSPhaseDefinition&
 MSPhasedTrafficLightLogic::getPhase(int givenStep) const {
-    assert(myPhases.size() > givenStep);
+    assert((int)myPhases.size() > givenStep);
     return *myPhases[givenStep];
 }
 
@@ -206,7 +206,7 @@ MSPhasedTrafficLightLogic::changeStepAndDuration(MSTLLogicControl& tlcontrol,
 /****************************************************************************/
 void
 MSPhasedTrafficLightLogic::setPhases(const Phases& phases, int step) {
-    assert(step < phases.size());
+    assert(step < (int)phases.size());
     deletePhases();
     myPhases = phases;
     myStep = step;
